requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev.txt requirements/dev.in
#
click==6.7                # via python-dotenv
dj-database-url==0.4.2
dj-email-url==0.0.10
django-configurations==2.0
Django==1.10.5
djangorestframework==3.5.3
honcho==0.7.1
psycopg2==2.6.2
python-dotenv==0.6.1
